demo/requirements.txt
Django<2.1
choice-enum==0.9
django-bootstrap3<10.1
Pillow==5.2.0
python-dateutil==2.7.3
django-extensions<2.1
feedparser==5.2.1
pycodestyle
-e ..
